Refactor FileContentType into a new file
